requirements.txt
altair==5.2.0
attrs==23.2.0
backports.zoneinfo==0.2.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
ctransformers==0.2.27
diskcache==5.6.3
elasticsearch==7.17.3
filelock==3.13.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
huggingface-hub==0.20.2
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
llama_cpp_python==0.2.27
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.0.3
pillow==10.2.0
pkgutil_resolve_name==1.3.10
protobuf==4.25.1
py-cpuinfo==9.0.0
pyarrow==14.0.2
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.16.2
safetensors==0.4.1
six==1.16.0
smmap==5.0.1
streamlit==1.29.0
sympy==1.12
tenacity==8.2.3
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.2
tornado==6.4
tqdm==4.66.1
transformers==4.36.2
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==1.26.18
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
